Fixing bug on delete Added test framework to POM

+++ services/ochestrator/src/main/java/com/graduationproject/ochestrator/saga/SagaParticipators/CreateEmployee.java
@@ -19,6 +19,7 @@ import java.util.UUID;
 
 import static com.graduationproject.ochestrator.topic.employee.EmployeeTopics.CreateEmployeeSagaBegin;
 import static com.graduationproject.ochestrator.topic.employee.EmployeeTopics.CreateEmployeeSagaFailed;
+import static java.util.Objects.nonNull;
 
 @Service
 public class CreateEmployee implements SagaParticipator<EmployeeDto> {
@@ -66,7 +67,7 @@ public class CreateEmployee implements SagaParticipator<EmployeeDto> {
         //this will be run after a successful saga
         Employee employee = employeeRepository.findEmployeeBySagaId(sagaId);
         employeeRepository.deleteBySagaId(sagaId);
-        if (employeeRepository.countByDepartment(employee.getDepartment()) == 0) {
+        if (nonNull(employee.getDepartment()) && employeeRepository.countByDepartment(employee.getDepartment()) == 0) {
             departmentRepository.deleteBySagaId(sagaId);
         }
